+++ OnlySnarf/__init__.py
@@ -1,2 +1,3 @@
+from .api import create_app
 from .snarf import Snarf
 from .menu import Menu