+++ bin/start-api-dev.sh
@@ -1,5 +1,2 @@
 #!/bin/sh
-# export FLASK_APP=./api/index.py
-# export FLASK_ENV=production
-# pipenv run flask --debug run -h 0.0.0.0
-python api/index.py -debug -verbose -verbose -verbose users
+python ./OnlySnarf/api/index.py -debug -verbose -verbose -verbose users