+++ bin/test-all.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+python -m pytest tests/api/test_api.py
+pytest tests/selenium
+pytest tests/snarf