+++ tests/snarf/test_message.py
@@ -10,35 +10,37 @@ from OnlySnarf.snarf import Snarf
 class TestSnarf(unittest.TestCase):
 
     def setUp(self):
-        config["input"] = ["/home/skeetzo/Projects/onlysnarf/public/images/shnarf.jpg", "/home/skeetzo/Projects/onlysnarf/public/images/snarf.jpg"]
-        config["price"] = DEFAULT.PRICE_MINIMUM
         config["text"] = "test balls"
         config["user"] = "random"
         # config["keep"] = True
         Settings.set_debug("tests")
+        # config["skip_download"] = False
+        # config["skip_upload"] = True
 
     def tearDown(self):
         config["input"] = []
         config["price"] = 0
-        config["user"] = None
+        # config["user"] = None
         # config["keep"] = False
         Snarf.close()
 
     def test_message(self):
-        assert Snarf.message(), "unable to send message"
+        assert Snarf.message(), "unable to send basic message"
 
-    # @unittest.skip("pointless")
-    # def test_message_files(self):
-    #     successful, results = Snarf.message()
-    #     assert Snarf.message(), "unable to upload message files"
+    def test_message_files_local(self):
+        config["input"] = ["/home/skeetzo/Projects/onlysnarf/public/images/shnarf.jpg", "/home/skeetzo/Projects/onlysnarf/public/images/snarf.jpg"]
+        successful = Snarf.message()
+        assert Snarf.message(), "unable to upload message files - local"
 
-    # @unittest.skip("pointless")
-    # def test_message_price(self):
-    #     assert Snarf.message(), "unable to set message price"
+    def test_message_files_remote(self):
+        config["input"] = ["https://github.com/skeetzo/onlysnarf/blob/master/public/images/shnarf.jpg?raw=true", "https://github.com/skeetzo/onlysnarf/blob/master/public/images/snarf.jpg?raw=true"]
+        successful = Snarf.message()
+        assert Snarf.message(), "unable to upload message files - remote"
 
-    # @unittest.skip("pointless")
-    # def test_message_text(self):
-    #     assert Snarf.message(), "unable to set message text"
+    def test_message_price(self):
+        config["input"] = ["/home/skeetzo/Projects/onlysnarf/public/images/shnarf.jpg"]
+        config["price"] = DEFAULT.PRICE_MINIMUM
+        assert Snarf.message(), "unable to set message price"
 
 ############################################################################################