+++ tests/snarf/test_post.py
@@ -13,7 +13,7 @@ class TestSnarf(unittest.TestCase):
         config["input"] = ["/home/skeetzo/Projects/onlysnarf/public/images/shnarf.jpg", "/home/skeetzo/Projects/onlysnarf/public/images/snarf.jpg"]
         config["price"] = DEFAULT.PRICE_MINIMUM
         config["text"] = "test balls"
-        config["tags"] = ["test","ticles"]
+        config["keywords"] = ["test","ticles"]
         config["performers"] = ["yourmom","yourdad"]
         Settings.set_debug("tests")
 
@@ -21,7 +21,7 @@ class TestSnarf(unittest.TestCase):
         config["input"] = []
         config["performers"] = []
         config["price"] = 0
-        config["tags"] = []
+        config["keywords"] = []
         Snarf.close()
 
     def test_post(self):